Update minio-go library
