+++ Godeps/Godeps.json
@@ -24,6 +24,7 @@
 		},
 		{
 			"ImportPath": "github.com/minio/minio-go",
+			"Comment": "v0.2.5-62-g61f6570",
 			"Rev": "61f6570da0edd761974216c9ed5da485d3cc0c99"
 		},
 		{
